refactor(internal): move each command to its own file
