+++ internal/flags/flags_test.go
@@ -2,7 +2,7 @@
  * Copyright (c) Intel Corporation 2021
  * SPDX-License-Identifier: Apache-2.0
  **********************************************************************/
-package rpc
+package flags
 
 import (
 	"errors"
