+++ internal/rps/rps_test.go
@@ -7,7 +7,7 @@ package rps
 import (
 	"net/http"
 	"net/http/httptest"
-	"rpc"
+	"rpc/internal/flags"
 	"strings"
 	"sync"
 	"testing"
@@ -38,30 +38,30 @@ func echo(w http.ResponseWriter, r *http.Request) {
 
 var testServer *httptest.Server
 var testUrl string
-var flags *rpc.Flags
+var testFlags *flags.Flags
 
 func init() {
 	// Create test server with the echo handler.
 	testServer = httptest.NewServer(http.HandlerFunc(echo))
 	// Convert http to ws
-	flags = rpc.NewFlags([]string{})
+	testFlags = flags.NewFlags([]string{})
 	testUrl = "ws" + strings.TrimPrefix(testServer.URL, "http")
-	flags.URL = testUrl
+	testFlags.URL = testUrl
 }
 
 func TestPrepareInitialMessage(t *testing.T) {
-	payload, payload1 := PrepareInitialMessage(flags)
+	payload, payload1 := PrepareInitialMessage(testFlags)
 	assert.NotEqual(t, payload, payload1)
 }
 
 func TestConnect(t *testing.T) {
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	err := server.Connect(true)
 	defer server.Close()
 	assert.NoError(t, err)
 }
 func TestSend(t *testing.T) {
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	err := server.Connect(true)
 	defer server.Close()
 	assert.NoError(t, err)
@@ -71,7 +71,7 @@ func TestSend(t *testing.T) {
 	server.Send(message)
 }
 func TestListen(t *testing.T) {
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	err := server.Connect(true)
 	defer server.Close()
 	assert.NoError(t, err)
@@ -96,7 +96,7 @@ func TestProcessMessageHeartbeat(t *testing.T) {
 	activation := `{
         "method": "heartbeat_request"
     }`
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	server.Connect(true)
 	decodedMessage := server.ProcessMessage([]byte(activation))
 	assert.NotNil(t, decodedMessage)
@@ -106,7 +106,7 @@ func TestProcessMessageSuccess(t *testing.T) {
         "method": "success",
         "message": "{\"status\":\"ok\", \"network\":\"configured\", \"ciraConnection\":\"configured\"}"
     }`
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	server.Connect(true)
 	decodedMessage := server.ProcessMessage([]byte(activation))
 	assert.Nil(t, decodedMessage)
@@ -116,7 +116,7 @@ func TestProcessMessageUnformattedSuccess(t *testing.T) {
         "method": "success",
         "message": "configured"
     }`
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	server.Connect(true)
 	decodedMessage := server.ProcessMessage([]byte(activation))
 	assert.Nil(t, decodedMessage)
@@ -126,7 +126,7 @@ func TestProcessMessageError(t *testing.T) {
         "method": "error",
         "message": "can't do it"
     }`
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	server.Connect(true)
 	decodedMessage := server.ProcessMessage([]byte(activation))
 	assert.Nil(t, decodedMessage)
@@ -137,7 +137,7 @@ func TestProcessMessageForLMS(t *testing.T) {
         "message": "ok",
         "payload": "eyJzdGF0dXMiOiJvayIsICJuZXR3b3JrIjoiY29uZmlndXJlZCIsICJjaXJhQ29ubmVjdGlvbiI6ImNvbmZpZ3VyZWQifQ=="
     }`
-	server := NewAMTActivationServer(flags)
+	server := NewAMTActivationServer(testFlags)
 	server.Connect(true)
 	decodedMessage := server.ProcessMessage([]byte(activation))
 	assert.Equal(t, []byte("{\"status\":\"ok\", \"network\":\"configured\", \"ciraConnection\":\"configured\"}"), decodedMessage)
